requirements.txt
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
connexion==2.9.0
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
google-api-core==2.2.2
google-auth==2.3.3
google-cloud==0.34.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.18.0
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
Werkzeug==1.0.1
